move Helper class into DatabaseRemove file
